Fix infinite loop in NewGameChooserModel.java when map download is corrupt.
Map data is loaded using a while loop: "while(entry != null)". The entry is incremented at the bottom of the loop: "entry = zis.getNextEntry()"

If the map data coming from zip file is truncated or corrupt, we can load a null url. Before we increment the entry, there is this statement "if(url == null) { continue; }". That skips the increment of the entry and we retry the same file without end.

Rather than just skip over the file, since the data is corrupt and the map data from it is highly questionable, the user is prompted if they would like to delete the file (so at least this error state can then be self-repaired for the next time we parse the maps).