Remove most of the custom thread pool implementation with ExecutorService.
For the most part our implementation of ThreadPool is replaced with ExecutorService. The reason for this is "waitForAll" expects the thread pool to still be usable afterwards. ExecutorService to blcok appropriately needs a shutdown() first and then is no longer usable. In the cases where we just want a straight ThreadPool, ExecutorService has been substituted for ThreadPool, otherwise the implmenetation of ThreadPool has been replaced.