Rename ui Util.java to SwingLib and increase usage of the class
Util.java main usage was to wrap SwingUtilities.invokeAndWait. To make the utility class and methods easier to find, they were renamed to SwingLib.invokeAndWait. After this was done, client code in the rest of the code base calling SwingUtilities was updated to call SwingLib, often simplifying removing a try catch block.

In addition, a chunk of code to request window focus was found in several places and ractored to the method: "SwingLib.requestWindowFocus(Component)"