Remove DownloadMapDialog, instead go directly to the InstallMapDialog screen
Optimize for the common use case of downloading maps from the main download site and go directly to that screen. When a user clicks download maps we first ask them to choose from a list of length 1 of download sites to use. Rather than have the extra click, let us just go to the download maps screen. In the future we can merge the two screens loading the default maps in a thread, allowing a user to select another site if they so choose.